requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.7.0
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langsmith==0.2.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
numpy==2.0.2
openai==1.57.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
packaging==24.2
propcache==0.2.1
protobuf==5.29.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.4.5
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
spotipy==2.24.0
SQLAlchemy==2.0.36
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tiktoken==0.8.0
tokenizers==0.20.4
tqdm==4.67.1
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3
